Add improvements to exception handling messages
Minor improvements to the exception service:
* split exception.catcher into a function per use case
* add documentation for recommended usage
* rename to exception.service.js to follow naming conventions

Add a guard clause to the `$exceptionHandler` decorator to ensure an
error object.

Add a missing `.catch` clause to the login function.
