--- client/app/blocks/exception/exception.js
@@ -1,20 +1,0 @@
-/** @ngInject */
-export function exception(logger) {
-  var service = {
-    catcher: catcher,
-  };
-
-  return service;
-
-  function catcher(message) {
-    return function(error) {
-      if (error.data && error.data.description) {
-        message += '\n' + error.data.description;
-        error.data.description = message;
-      }
-      logger.error(message, error);
-
-      return Promise.reject(error);
-    };
-  }
-}
